Separate into example file and module
